+++ CHANGELOG.md
@@ -6,7 +6,8 @@ ProofNetworkJdf Changelog
  * --
 
 #### Bug Fixing
- * --
+ * XJDF: Fix missing namespaces (for updateFrom)
+ * XJDF: Fix read and write DeviceID attribute
 
 ## 0.19.8.7
 #### Features